+++ src/plugins/chess960/__init__.py
@@ -1,0 +1,6 @@
+from pathlib import Path
+
+from plugins import PLUGINS_DIR
+
+PLUGIN_NAME: str = 'chess960'
+PLUGIN_DIR: Path = PLUGINS_DIR / PLUGIN_NAME